Add files via upload
This project is of pure php, html, css and javascript which involves users signing up  and registering their names and other details which helps them in booking a car parking slot. Note that this project employs the concept of users signing up their details and registering their details then the details are solved in a database called assignment.